+++ app/routes/components/ScrollArea/demos/WithAbsolutePosition/index.tsx
@@ -1,6 +1,7 @@
+import { useTheme } from "@/components/CopyUiProvider";
 import { Flex } from "@/components/Flex";
 import { ScrollArea } from "@/components/ScrollArea";
-import { useTheme } from "@/components/ThemeProvider";
+
 import { lorem } from "./lorem";
 
 export default function Demo() {
@@ -8,7 +9,7 @@ export default function Demo() {
 
   const containerStyle = {
     width: "30rem",
-    height: "30rem",
+    height: "20rem",
   };
 
   const scrollAreaStyle = {
@@ -35,9 +36,7 @@ export default function Demo() {
               <p>{lorem}</p>
             </ScrollArea.Content>
           </ScrollArea.Viewport>
-          <ScrollArea.Scrollbar>
-            <ScrollArea.Thumb />
-          </ScrollArea.Scrollbar>
+          <ScrollArea.ScrollbarWithThumb />
         </ScrollArea>
       ))}
     </Flex>