+++ src/components/ScrollArea/CHANGELOG.md
@@ -7,3 +7,5 @@
 - 2025-10-30: Add the variant `absolute` (it uses absolute positioning to make
   viewport just fill the container).
 - 2025-11-04: Support the `ScrollArea.ScrollbarWithThumb`.
+- 2025-12-11: Support the horizontal scrollbars with prop
+  `orientation="horizontal"`.